+++ examples/simulated/plot_classifier_comparison.py
@@ -27,7 +27,7 @@ from matplotlib.colors import ListedColormap
 from sklearn.model_selection import train_test_split
 
 from pyriemann.datasets import make_covariances, make_gaussian_blobs
-from pyriemann.classification import MDM, KNearestNeighbor
+from pyriemann.classification import MDM, KNearestNeighbor, SVC
 
 
 ###############################################################################
@@ -45,11 +45,13 @@ def get_proba(cov_00, cov_01, cov_11, clf):
 
 names = [
     "MDM",
-    "Nearest Neighbors"
+    "Nearest Neighbors",
+    "SVC"
 ]
 classifiers = [
     MDM(),
-    KNearestNeighbor(n_neighbors=3)
+    KNearestNeighbor(n_neighbors=3),
+    SVC(probability=True)
 ]
 n_classifs = len(classifiers)
 
@@ -96,7 +98,7 @@ cm_bright = ListedColormap(["#FF0000", "#0000FF"])
 
 ###############################################################################
 
-figure = plt.figure(figsize=(8, 10))
+figure = plt.figure(figsize=(12, 10))
 i = 1
 # iterate over datasets
 for ds_cnt, (X, y) in enumerate(datasets):