+++ pyriemann/classification.py
@@ -1,21 +1,25 @@
 """Module for classification function."""
+import functools
+
 import numpy as np
 
 from scipy import stats
 
 from sklearn.base import BaseEstimator, ClassifierMixin, TransformerMixin
+from sklearn.svm import SVC as sklearnSVC
 from sklearn.utils.extmath import softmax
 from sklearn.linear_model import LogisticRegression
 from sklearn.pipeline import make_pipeline
+
 from joblib import Parallel, delayed
 
+from .utils.kernel import kernel
 from .utils.mean import mean_covariance
 from .utils.distance import distance
 from .tangentspace import FGDA, TangentSpace
 
 
 class MDM(BaseEstimator, ClassifierMixin, TransformerMixin):
-
     """Classification by Minimum Distance to Mean.
 
     Classification by nearest centroid. For each of the given classes, a
@@ -24,7 +28,7 @@ class MDM(BaseEstimator, ClassifierMixin, TransformerMixin):
 
     Parameters
     ----------
-    metric : string | dict (default: 'riemann')
+    metric : string | dict, default: 'riemann'
         The type of metric used for centroid and distance estimation.
         see `mean_covariance` for the list of supported metric.
         the metric could be a dict with two keys, `mean` and `distance` in
@@ -32,7 +36,7 @@ class MDM(BaseEstimator, ClassifierMixin, TransformerMixin):
         distance estimation. Typical usecase is to pass 'logeuclid' metric for
         the mean in order to boost the computional speed and 'riemann' for the
         distance in order to keep the good sensitivity for the classification.
-    n_jobs : int, (default: 1)
+    n_jobs : int, default: 1
         The number of jobs to use for the computation. This works by computing
         each of the class centroid in parallel.
         If -1 all CPUs are used. If 1 is given, no parallel computing code is
@@ -92,12 +96,12 @@ class MDM(BaseEstimator, ClassifierMixin, TransformerMixin):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
-        y : ndarray shape (n_trials, 1)
-            labels corresponding to each trial.
-        sample_weight : None | ndarray shape (n_trials, 1)
-            the weights of each sample. if None, each sample is treated with
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : ndarray, shape (n_matrices,)
+            Labels corresponding to each matrix.
+        sample_weight : None | ndarray shape (n_matrices,)
+            Weights of each matrix. If None, each matrix is treated with
             equal weights.
 
         Returns
@@ -139,32 +143,32 @@ class MDM(BaseEstimator, ClassifierMixin, TransformerMixin):
         return dist
 
     def predict(self, covtest):
-        """get the predictions.
+        """Get the predictions.
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
 
         Returns
         -------
-        pred : ndarray of int, shape (n_trials, 1)
-            the prediction for each trials according to the closest centroid.
+        pred : ndarray of int, shape (n_matrices,)
+            Predictions for each matrix according to the closest centroid.
         """
         dist = self._predict_distances(covtest)
         return self.classes_[dist.argmin(axis=1)]
 
     def transform(self, X):
-        """get the distance to each centroid.
+        """Get the distance to each centroid.
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
 
         Returns
         -------
-        dist : ndarray, shape (n_trials, n_classes)
+        dist : ndarray, shape (n_matrices, n_classes)
             the distance to each centroid according to the metric.
         """
         return self._predict_distances(X)
@@ -179,19 +183,18 @@ class MDM(BaseEstimator, ClassifierMixin, TransformerMixin):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
 
         Returns
         -------
-        prob : ndarray, shape (n_trials, n_classes)
-            the softmax probabilities for each class.
+        prob : ndarray, shape (n_matrices, n_classes)
+            Probabilities for each class.
         """
         return softmax(-self._predict_distances(X)**2)
 
 
 class FgMDM(BaseEstimator, ClassifierMixin, TransformerMixin):
-
     """Classification by Minimum Distance to Mean with geodesic filtering.
 
     Apply geodesic filtering described in [1]_, and classify using MDM.
@@ -202,7 +205,7 @@ class FgMDM(BaseEstimator, ClassifierMixin, TransformerMixin):
 
     Parameters
     ----------
-    metric : string | dict (default: 'riemann')
+    metric : string | dict, default: 'riemann'
         The type of metric used for centroid and distance estimation.
         see `mean_covariance` for the list of supported metric.
         the metric could be a dict with two keys, `mean` and `distance` in
@@ -210,12 +213,12 @@ class FgMDM(BaseEstimator, ClassifierMixin, TransformerMixin):
         distance estimation. Typical usecase is to pass 'logeuclid' metric for
         the mean in order to boost the computional speed and 'riemann' for the
         distance in order to keep the good sensitivity for the classification.
-    tsupdate : bool (default False)
+    tsupdate : bool, default False
         Activate tangent space update for covariante shift correction between
         training and test, as described in [2]_. This is not compatible with
         online implementation. Performance are better when the number of trials
         for prediction is higher.
-    n_jobs : int, (default: 1)
+    n_jobs : int, default: 1
         The number of jobs to use for the computation. This works by computing
         each of the class centroid in parallel.
         If -1 all CPUs are used. If 1 is given, no parallel computing code is
@@ -271,10 +274,10 @@ class FgMDM(BaseEstimator, ClassifierMixin, TransformerMixin):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
-        y : ndarray shape (n_trials, 1)
-            labels corresponding to each trial.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : ndarray, shape (n_matrices,)
+            Labels corresponding to each matrix.
 
         Returns
         -------
@@ -290,17 +293,17 @@ class FgMDM(BaseEstimator, ClassifierMixin, TransformerMixin):
         return self
 
     def predict(self, X):
-        """get the predictions after FGDA filtering.
+        """Get the predictions after FGDA filtering.
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
 
         Returns
         -------
-        pred : ndarray of int, shape (n_trials, 1)
-            the prediction for each trials according to the closest centroid.
+        pred : ndarray of int, shape (n_matrices,)
+            Predictions for each matrix according to the closest centroid.
         """
         cov = self._fgda.transform(X)
         return self._mdm.predict(cov)
@@ -310,28 +313,28 @@ class FgMDM(BaseEstimator, ClassifierMixin, TransformerMixin):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
 
         Returns
         -------
-        prob : ndarray, shape (n_trials, n_classes)
+        prob : ndarray, shape (n_matrices, n_classes)
             the softmax probabilities for each class.
         """
         cov = self._fgda.transform(X)
         return self._mdm.predict_proba(cov)
 
     def transform(self, X):
-        """get the distance to each centroid after FGDA filtering.
+        """Get the distance to each centroid after FGDA filtering.
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
 
         Returns
         -------
-        dist : ndarray, shape (n_trials, n_cluster)
+        dist : ndarray, shape (n_matrices, n_cluster)
             the distance to each centroid according to the metric.
         """
         cov = self._fgda.transform(X)
@@ -339,7 +342,6 @@ class FgMDM(BaseEstimator, ClassifierMixin, TransformerMixin):
 
 
 class TSclassifier(BaseEstimator, ClassifierMixin):
-
     """Classification in the tangent space.
 
     Project data in the tangent space and apply a classifier on the projected
@@ -348,7 +350,7 @@ class TSclassifier(BaseEstimator, ClassifierMixin):
 
     Parameters
     ----------
-    metric : string | dict (default: 'riemann')
+    metric : string | dict, default: 'riemann'
         The type of metric used for centroid and distance estimation.
         see `mean_covariance` for the list of supported metric.
         the metric could be a dict with two keys, `mean` and `distance` in
@@ -356,12 +358,12 @@ class TSclassifier(BaseEstimator, ClassifierMixin):
         distance estimation. Typical usecase is to pass 'logeuclid' metric for
         the mean in order to boost the computional speed and 'riemann' for the
         distance in order to keep the good sensitivity for the classification.
-    tsupdate : bool (default False)
+    tsupdate : bool, default: False
         Activate tangent space update for covariante shift correction between
         training and test, as described in [2]. This is not compatible with
         online implementation. Performance are better when the number of trials
         for prediction is higher.
-    clf: sklearn classifier (default LogisticRegression)
+    clf: sklearn classifier, default: LogisticRegression
         The classifier to apply in the tangent space
 
     See Also
@@ -388,10 +390,10 @@ class TSclassifier(BaseEstimator, ClassifierMixin):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
-        y : ndarray shape (n_trials, 1)
-            labels corresponding to each trial.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : ndarray, shape (n_matrices,)
+            Labels corresponding to each matrix.
 
         Returns
         -------
@@ -405,38 +407,37 @@ class TSclassifier(BaseEstimator, ClassifierMixin):
         return self
 
     def predict(self, X):
-        """get the predictions.
+        """Get the predictions.
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
 
         Returns
         -------
-        pred : ndarray of int, shape (n_trials, 1)
-            the prediction for each trials according to the closest centroid.
+        pred : ndarray of int, shape (n_matrices,)
+            Predictions for each matrix according to the closest centroid.
         """
         return self._pipe.predict(X)
 
     def predict_proba(self, X):
-        """get the probability.
+        """Get the probability.
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
 
         Returns
         -------
-        pred : ndarray of ifloat, shape (n_trials, n_classes)
-            the prediction for each trials according to the closest centroid.
+        pred : ndarray of ifloat, shape (n_matrices, n_classes)
+            Predictions for each matrix according to the closest centroid.
         """
         return self._pipe.predict_proba(X)
 
 
 class KNearestNeighbor(MDM):
-
     """Classification by K-NearestNeighbor.
 
     Classification by nearest Neighbors. For each point of the test set, the
@@ -446,12 +447,12 @@ class KNearestNeighbor(MDM):
 
     Parameters
     ----------
-    n_neighbors : int, (default: 5)
+    n_neighbors : int, default: 5
         Number of neighbors.
-    metric : string | dict (default: 'riemann')
+    metric : string | dict, default: 'riemann'
         The type of metric used for distance estimation.
         see `distance` for the list of supported metric.
-    n_jobs : int, (default: 1)
+    n_jobs : int, default: 1
         The number of jobs to use for the computation. This works by computing
         each of the distance to the training set in parallel.
         If -1 all CPUs are used. If 1 is given, no parallel computing code is
@@ -486,10 +487,10 @@ class KNearestNeighbor(MDM):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
-        y : ndarray shape (n_trials, 1)
-            labels corresponding to each trial.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : ndarray, shape (n_matrices,)
+            Labels corresponding to each matrix.
 
         Returns
         -------
@@ -503,17 +504,17 @@ class KNearestNeighbor(MDM):
         return self
 
     def predict(self, covtest):
-        """get the predictions.
+        """Get the predictions.
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
 
         Returns
         -------
-        pred : ndarray of int, shape (n_trials, 1)
-            the prediction for each trials according to the closest centroid.
+        pred : ndarray of int, shape (n_matrices,)
+            Predictions for each matrix according to the closest centroid.
         """
         dist = self._predict_distances(covtest)
         neighbors_classes = self.classmeans_[np.argsort(dist)]
@@ -549,3 +550,167 @@ class KNearestNeighbor(MDM):
                 )
 
         return prob
+
+
+class SVC(sklearnSVC):
+    """Classification by Riemannian Support Vector Machine.
+
+    Support vector machine with precomputed Riemannian kernel matrix
+    according to different metrics as described in [1]_.
+
+    Parameters
+    ----------
+    metric : {'riemann', 'euclid', 'logeuclid'}, default: 'riemann'
+        Metric for kernel matrix computation.
+    Cref : None | ndarray | callable, shape (n_channels, n_channels)
+        Reference point for kernel matrix computation.
+        If None, the mean of the training data according to the metric is used.
+        If callable, the function is called on the training data to calculate
+        Cref.
+    kernel_fct : 'precomputed' | callable
+        If 'precomputed', the kernel matrix for datasets X and Y is estimated
+        according to pyriemann.utils.kernel(X, Y, Cref, metric).
+        If callable, the callable is passed as the kernel parameter to
+        sklearn.svm.SVC(). The callable has to be of the form
+        kernel(X, Y, Cref, metric).
+    C : float, default: 1.0
+        Regularization parameter. The strength of the regularization is
+        inversely proportional to C. Must be strictly positive. The penalty
+        is a squared l2 penalty.
+    shrinking : bool, default: True
+        Whether to use the shrinking heuristic.
+    probability : bool, default: False
+        Whether to enable probability estimates. This must be enabled prior
+        to calling `fit`, will slow down that method as it internally uses
+        5-fold cross-validation, and `predict_proba` may be inconsistent with
+        `predict`. Read more in the :ref:`User Guide <scores_probabilities>`.
+    tol : float, default: 1e-3
+        Tolerance for stopping criterion.
+    cache_size : float, default: 200
+        Specify the size of the kernel cache (in MB).
+    class_weight : dict or 'balanced', default: None
+        Set the parameter C of class i to class_weight[i]*C for
+        SVC. If not given, all classes are supposed to have
+        weight one.
+        The "balanced" mode uses the values of y to automatically adjust
+        weights inversely proportional to class frequencies in the input data
+        as ``n_matrices / (n_classes * np.bincount(y))``.
+    verbose : bool, default: False
+        Enable verbose output. Note that this setting takes advantage of a
+        per-process runtime setting in libsvm that, if enabled, may not work
+        properly in a multithreaded context.
+    max_iter : int, default: -1
+        Hard limit on iterations within solver, or -1 for no limit.
+    decision_function_shape : {'ovo', 'ovr'}, default: 'ovr'
+        Whether to return a one-vs-rest ('ovr') decision function of shape
+        (n_matrices, n_classes) as all other classifiers, or the original
+        one-vs-one ('ovo') decision function of libsvm which has shape
+        (n_matrices, n_classes * (n_classes - 1) / 2). However, note that
+        internally, one-vs-one ('ovo') is always used as a multi-class strategy
+        to train models; an ovr matrix is only constructed from the ovo matrix.
+        The parameter is ignored for binary classification.
+    break_ties : bool, default: False
+        If true, ``decision_function_shape='ovr'``, and number of classes > 2,
+        :term:`predict` will break ties according to the confidence values of
+        :term:`decision_function`; otherwise the first class among the tied
+        classes is returned. Please note that breaking ties comes at a
+        relatively high computational cost compared to a simple predict.
+    random_state : int, RandomState instance or None, default: None
+        Controls the pseudo random number generation for shuffling the data for
+        probability estimates. Ignored when `probability` is False.
+        Pass an int for reproducible output across multiple function calls.
+        See :term:`Glossary <random_state>`.
+
+    Notes
+    -----
+    .. versionadded:: 0.2.8
+
+    References
+    ----------
+    .. [1] A. Barachant, S. Bonnet, M. Congedo, and C. Jutten.
+        Classification of covariance matrices using a Riemannian-based kernel
+        for BCI applications". In: Neurocomputing 112 (July 2013), pp. 172-178.
+    """
+
+    def __init__(self,
+                 *,
+                 metric='riemann',
+                 kernel_fct=None,
+                 Cref=None,
+                 C=1.0,
+                 shrinking=True,
+                 probability=False,
+                 tol=1e-3,
+                 cache_size=200,
+                 class_weight=None,
+                 verbose=False,
+                 max_iter=-1,
+                 decision_function_shape="ovr",
+                 break_ties=False,
+                 random_state=None):
+        """Init."""
+        self.Cref = Cref
+        self.metric = metric
+        self.Cref_ = None
+        self.kernel_fct = kernel_fct
+        super().__init__(kernel='precomputed',
+                         C=C,
+                         shrinking=shrinking,
+                         probability=probability,
+                         tol=tol,
+                         cache_size=cache_size,
+                         class_weight=class_weight,
+                         verbose=verbose,
+                         max_iter=max_iter,
+                         decision_function_shape=decision_function_shape,
+                         break_ties=break_ties,
+                         random_state=random_state
+                         )
+
+    def fit(self, X, y, sample_weight=None):
+        """Fit.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : ndarray, shape (n_matrices,)
+            Labels corresponding to each matrix.
+        sample_weight : ndarray, shape (n_matrices,), default: None
+            Per-sample weights. Rescale C per sample. Higher weights
+            force the classifier to put more emphasis on these points.
+
+        Returns
+        -------
+        self : Riemannian SVC instance
+            The SVC instance.
+        """
+        self._set_cref(X)
+        self._set_kernel()
+        super().fit(X, y)
+        return self
+
+    def _set_cref(self, X):
+        if self.Cref is None:
+            self.Cref_ = mean_covariance(X, metric=self.metric)
+        elif callable(self.Cref):
+            self.Cref_ = self.Cref(X)
+        elif isinstance(self.Cref, np.ndarray):
+            self.Cref_ = self.Cref
+        else:
+            raise TypeError(f'Cref must be np.ndarray, callable or None, is'
+                            f' {self.Cref}.')
+
+    def _set_kernel(self):
+        if callable(self.kernel_fct):
+            self.kernel = functools.partial(self.kernel_fct,
+                                            Cref=self.Cref_,
+                                            metric=self.metric)
+
+        elif self.kernel_fct is None:
+            self.kernel = functools.partial(kernel,
+                                            Cref=self.Cref_,
+                                            metric=self.metric)
+        else:
+            raise TypeError(f"kernel must be 'precomputed' or callable, is "
+                            f"{self.kernel}.")