+++ doc/whatsnew.rst
@@ -52,6 +52,13 @@ v0.2.8.dev
 
 - Add ``predict_proba`` in :class:`pyriemann.classification.KNearestNeighbor`, and correct attribute ``classes_``
 
+- Add Riemannian Support Vector Machine Classification: :class:`pyriemann.classification.SVC`
+
+- Add Riemannian Support Vector Machine Regression: :class:`pyriemann.regression.SVR`
+
+- Add K-Nearest-Neighbor Regression: :class:`pyriemann.regression.KNearestNeighborRegressor`
+
+- Add example to compare classifiers on synthetic datasets
 
 v0.2.7 (June 2021)
 ------------------